move "getDB" from rmgpy.rmg.model to rmgpy.data.rmg module
