+++ rmgpy/data/rmg.py
@@ -42,7 +42,7 @@ from rmgpy.data.kinetics.database import KineticsDatabase
 from statmech import StatmechDatabase
 from solvation import SolvationDatabase
 
-from rmgpy.scoop_framework.util import broadcast
+from rmgpy.scoop_framework.util import get, broadcast
 
 # Module-level variable to store the (only) instance of RMGDatabase in use.
 database = None
@@ -226,3 +226,41 @@ class RMGDatabase:
         self.forbiddenStructures.saveOld(os.path.join(path, 'ForbiddenStructures.txt'))
         self.kinetics.saveOld(path)
         self.statmech.saveOld(path)
+
+def getDB(name):
+    """
+    Returns the RMG database object that corresponds
+    to the parameter name.
+
+    First, the module level is queried. If this variable
+    is empty, the broadcasted variables are queried.
+    """
+    global database
+
+    if database:
+        if name == 'kinetics':
+            return database.kinetics
+        elif name == 'thermo':
+            return database.thermo
+        elif name == 'transport':
+            return database.transport
+        elif name == 'solvation':
+            return database.solvation
+        elif name == 'statmech':
+            return database.statmech
+        elif name == 'forbidden':
+            return database.forbiddenStructures
+        else:
+            raise Exception('Unrecognized database keyword: {}'.format(name))
+    else:
+        try:
+            db = get(name)
+            if db:
+                return db
+            else:
+                raise Exception
+        except Exception, e:
+            logging.error("Did not find a way to obtain the broadcasted database for {}.".format(name))
+            raise e
+
+    raise Exception('Could not get database with name: {}'.format(name))
